+++ js/util.js
@@ -53,4 +53,12 @@ const showAlert = (message) => {
   }, ALERT_SHOW_TIME);
 };
 
-export {getRandomInteger, createRandomIdFromRangeGenerator, createCommentId, isEscapeKey, isEnterKey, showAlert};
+function debounce (callback, timeoutDelay = 500) {
+  let timeoutId;
+  return (...rest) => {
+    clearTimeout(timeoutId);
+    timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
+  };
+}
+
+export {getRandomInteger, createRandomIdFromRangeGenerator, createCommentId, isEscapeKey, isEnterKey, showAlert, debounce};
